restructured test helpers
